# backend/api-gateway/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LumenHorizon Public API
  version: 0.1.0
  description: API Gateway contract. Public read routes are anonymous; admin routes require JWT bearer auth.
servers:
  - url: https://api.lumenhorizon.com
    description: Public API endpoint
  - url: http://localhost:8080
    description: Local API Gateway development
security: []
...
paths:
...
  /api/v1/admin/ingest/trigger:
    post:
      tags: [Admin]
      summary: Trigger ingest through an authorized admin operation
      operationId: triggerIngest
      security:
        - AdminJwt: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '202':
          $ref: '#/components/responses/AdminActionAccepted'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/UpstreamError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    AdminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT validated against configured issuer, audience, JWKS URL, required subject, expiry, and admin role claim.
...
  responses:
...
    AdminActionAccepted:
      description: Admin action accepted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated subject lacks permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UpstreamError:
      description: Internal dependency failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: Required dependency unavailable or service not ready
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, meta, error]
      additionalProperties: false
      properties:
        data: true
        meta:
          $ref: '#/components/schemas/EnvelopeMeta'
        error:
          type: 'null'
    ErrorEnvelope:
      type: object
      required: [data, meta, error]
      additionalProperties: false
      properties:
        data:
          type: 'null'
        meta:
          $ref: '#/components/schemas/EnvelopeMeta'
        error:
          $ref: '#/components/schemas/ErrorBody'
    EnvelopeMeta:
      type: object
      required: [request_id, timestamp]
      additionalProperties: false
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorBody:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - unauthenticated
            - forbidden
            - not_found
            - tile_not_found
            - conflict
            - unprocessable_entity
            - rate_limited
            - internal_error
            - upstream_error
            - service_unavailable
            - tile_unavailable
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
...